refactor: extract cardEdit to submodule #37
