+++ src/Components/Common/Card/EditCard/EditCard.style.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+import { CardContainer, rotate, rotateInverse } from '../Card.style';
+
+export const CardEditContainer = styled(CardContainer)`
+  animation-name: ${(props) => {
+    return props.active ? rotateInverse : rotate;
+  }};
+`;
