+++ src/Components/Common/Card/EditCard/index.ts
@@ -1,0 +1,1 @@
+export * from './EditCard';
